// src/types/modelGame.ts
export type GameData = {
    idGame: number;
    name: string;
    author?: string;
    publisher?: string;
    numberOfPlayers?: string;
    minAge?: string;
    duration?: string;
    type?: string;
    instructionLink?: string;
    playArea?: string;
    volunteerArea?: string;
    idZone?: number;
    toAnimate?: boolean;
    received?: boolean;
    mechanisms?: string;
    themes?: string;
    tags?: string;
    description?: string;
    imageUrl?: string;
    logoUrl?: string;
    videoUrl?: string;
};